template

+++ src/hotspot/share/utilities/spinYield.cpp
@@ -66,7 +66,7 @@ void SpinYield::report(outputStream* s) const {
   if (_sleep_time.value() != 0) { // Report sleep duration, if slept.
     separator = print_separator(s, separator);
     s->print("sleep = " UINT64_FORMAT " usecs",
-             (uint64_t)_sleep_time.microseconds());
+             _sleep_time.microseconds());
   }
   if (separator == initial_separator) {
     s->print("no waiting");
